+++ CHANGELOG.rst
@@ -58,6 +58,7 @@ Fixed
 - Make failover logging more verbose.
 - Fix hotreload for roles who leave gaps in httpd routes.
 - Check user e-mail uniquness when editing.
+- Expelled instances are removed from ``_cluster`` space.
 
 -------------------------------------------------------------------------------
 [2.6.0] - 2021-04-26